requirements.txt
google-auth>=2.23.0
google-api-python-client>=2.100.0
openai>=1.0.0
python-dotenv>=1.0.0
flask>=2.3.0
simple-salesforce>=1.12.0
fastmcp>=0.1.0
starlette>=0.27.0
requests>=2.31.0

